1) Global exception handler is added 2) Services is extended from CrudService

+++ src/main/java/com/amin/domain/repository/DepartmentRepository.java
@@ -4,5 +4,5 @@ import com.amin.domain.entity.Department;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface DepartmentRepository extends JpaRepository<Department, Integer> {
+public interface DepartmentRepository extends JpaRepository<Department, Long> {
 }
