+++ src/main/java/com/amin/domain/repository/DependentRepository.java
@@ -3,5 +3,5 @@ package com.amin.domain.repository;
 import com.amin.domain.entity.Dependent;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DependentRepository extends JpaRepository<Dependent, Integer> {
+public interface DependentRepository extends JpaRepository<Dependent, Long> {
 }
