+++ src/main/java/com/amin/domain/repository/JobRepository.java
@@ -3,5 +3,5 @@ package com.amin.domain.repository;
 import com.amin.domain.entity.Job;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface JobRepository extends JpaRepository<Job, Integer> {
+public interface JobRepository extends JpaRepository<Job, Long> {
 }
